api/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.2
boto3==1.34.16
botocore==1.34.16
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.7
deprecation==2.1.0
Django==5.0
django-storages==1.14.2
djangorestframework==3.14.0
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.12.2
gotrue==2.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
imageai==3.0.3
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mock==4.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pluggy==1.3.0
postgrest==0.13.1
py==1.11.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
pytest==7.1.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
realtime==1.0.2
requests==2.31.0
s3transfer==0.10.0
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.1
supafunc==0.3.3
sympy==1.12
tomli==2.0.1
torch==2.1.2+cpu
torchvision==0.16.2+cpu
tqdm==4.64.1
typing_extensions==4.9.0
urllib3==2.0.7
websockets==11.0.3
